Fix Fiends never going into their pain animation
There was a stray "return" statement in demon1_pain() which meant that
the function would always return without causing the monster to go into
its pain animation.  It looks like the "return" statement was added by
accident.  This commit removes it.

Fixing this also eliminates a compiler warning (FTEQCC previously issued
a warning about the fact that the code after the "return" statement was
unreachable).