Address warnings about th_pain() type mismatches
This commit irons out an issue that progs_dump inherited from id's
original code and which caused FTEQCC to issue "type mismatch" warnings.

The th_pain() field is defined (in defs.qc) as being a function that
takes two parameters, an entity and a float, and returns void.  However,
there were places in id's code where a function which didn't match this
signature was assigned to an entity's th_pain().

The problematic functions were:

- player_pain()
- dog_pain()
- shalrath_pain()
- nopain()
- fd_secret_use()
- SUB_Null()

In all cases, the problem was that these functions take no parameters,
whereas th_pain() is supposed to take two parameters (entity and float).

This problem could be solved in different ways.  In this commit, I've
used the same general approach that Lunaran uses in his Copper mod,
which is to leave the definition of th_pain() as it is, and modify the
rest of the code to fit that definition.

Fixing player_pain(), dog_pain(), shalrath_pain(), and nopain() is
straightforward.  This commit simply makes these functions take two
parameters (entity and float) so that they match th_pain().

The parameters of fd_secret_use() and SUB_Null() cannot be changed
because those functions are used for other purposes elsewhere in the
code.  This commit works around this by adding two trivial functions,
fd_secret_pain() and SUB_NullPain(), which match the signature of
th_pain() and which are assigned instead of the old functions.