Address warnings about "!" and parentheses
This commit fixes nine lines that caused FTEQCC to issue the following
warning:

    warning F304: suggest parenthesis for unary operator that applies to
    multiple terms

Which is a bit cryptic, unless the user already knows what it's getting
at.  What it's recommending is to change statements like this:

    if (!someflags & SPECIFIC_FLAG)

... into this:

    if (!(someflags & SPECIFIC_FLAG))

The reason the warning gets issued is the weird precedence that the "!"
operator has in QuakeC compared to other programming languages.  In
other languages, like C, the following line:

    if (!someflags & SPECIFIC_FLAG)

... would mean this:

    if ((!someflags) & (SPECIFIC_FLAG))

... whereas in QuakeC, it means this:

    if (!(someflags & SPECIFIC_FLAG))

So, adding the parentheses makes the meaning explicit.