Address a warning about a sound precache
There were a couple of lines in the SightSound() function which would,
in theory, play a sound for the nonexistent "Vomitus" monster, which id
never finished implementing.  However, those lines caused FTEQCC to
issue a warning because the sound file is never precached.  This commit
fixes this by removing those lines.