Fix gravity/ladder effects not reset on respawn
Previously, the new fields that trigger_setgravity and trigger_ladder
set on the player (gravity, wantedgravity, and onladder) were not reset
when the player respawns in deathmatch or coop.  In particular, this
meant that if the player touched a trigger_setgravity, then died and
respawned, they would end up with (for example) low gravity in a part of
the map which wasn't supposed to have low gravity.

This commit resets these three fields when the player spawns/respawns.