Fix players "sliding down" ladders due to gravity
Previously, whenever QuakeGuy was holding onto a ladder (but not
climbing), he would slowly slide down the ladder.  This caught my
attention because it doesn't happen in Rubicon 2, from which the ladder
code originates.

Compared to the Rubicon 2 code, the difference was that the
ladder-related code in ladder_touch() and PlayerPreThink() didn't set
the player's "gravity" field: the code instead set a new field,
"wantedgravity".  The trouble was that setting "wantedgravity" had no
effect: it wasn't checked and acted upon elsewhere in the code, and
unlike the "gravity" field, "wantedgravity" does not have special
meaning to the engine.

But certainly, in progs_dump, ladder_touch() mustn't set the player's
gravity directly, because then trigger_ladder and trigger_setgravity
would conflict if they were both used in the same map.

This commit implements the following solution:

- The player's gravity is not directly set by either ladder_touch() or
  trigger_gravity_touch().

- onladder keeps track of whether the player touched a ladder since the
  previous frame (same as before).

- wantedgravity is now used exclusively to keep track of the gravity
  setting from the last trigger_setgravity the player touched.  If the
  player hasn't touched one, it just holds the default gravity setting.

- In PlayerPreThink(), if self.onladder is true, then the player's
  gravity is disabled (like in Rubicon 2), otherwise self.wantedgravity
  is applied.

This should fix the player-sliding-down-ladder bug while allowing
trigger_ladder and trigger_setgravity to coexist peacefully.