Fix items not respawning in deathmatch mode
Item respawning was unfortunately broken in deathmatch mode: the only
things that would respawn were weapons, and the pentagram and ring
powerups.

It seemed like most of the old deathmatch item respawn logic had been
replaced by the new item respawn logic, which allows maps to set up
respawnable items in single-player/coop.  However, the new code mostly
didn't include logic to handle deathmatch mode.

This commit reinstates the original item respawn logic for deathmatch
mode.  It doesn't make any changes to the new logic that applies to
single-player/coop.

The only intentional change from the original deathmatch behavior is the
respawn logic for the megahealth, which had to be changed to fit in with
the new megahealth rot behavior; the comment I've added to
health_touch() explains this.