Prevent players "sticking" to the tops of ladders
One quirk of the ladder code from Rubicon 2 was that if the top of
a trigger_ladder was flush with the platform at the top of the ladder,
and if a player stood on that platform with the bottom of their bounding
box flush with the top of the trigger, the player would "stick" to the
ladder as if they were holding onto it.

It could be argued that a map just shouldn't put the top of
a trigger_ladder flush with a platform, but some released maps already
do this (jamx_bloodshot, jamx_jcr, jamx_kalebclark, and possibly
others), and it seems like something mappers are likely to do.

The effect of the player "sticking" to a ladder was also much more
noticeable after the previous commit, which fixed the ladder gravity
issue, so I felt something needed to be done about this.

This commit adds a check to ladder_touch() which makes it ignore
a player who is standing flush with the top of the trigger.