Gracefully handle trigger-spawn bugs in maps
When a mod supports trigger-spawned monsters, a common map bug is that
a map includes a trigger-spawned monster which can never be activated,
either because it has no targetname or because its targetname isn't
targeted by any other entity.  If left unchecked, this is a problem
because it makes it impossible for the player to get 100% kills.

Out of the progs_dump maps already released, two of the sm188 maps
(sm188_ing and sm188_jcr) have this bug.

This commit makes the progs automatically detect and handle this type of
map bug.  It adds in a check to see whether a trigger-spawned monster
has a targetname and whether that targetname is targeted by another
entity.  If this check fails, then a developer warning will be logged
and the monster will be removed, before it messes up the kill count.