Fix pain_target never firing in some cases
It was possible for most types of monster to die without firing their
pain_target.  This affected all monsters except Scrags, Shamblers,
Spawns, and Vores.

This could happen because most of the monsters' th_die() functions
called monster_pain_use() only in the case where the monster got gibbed,
and not in the case where the monster died normally.  As a result, if
the monster's health was above its pain_threshold, and the monster then
took enough damage in one go to kill it, but not gib it, the monster
wouldn't fire its pain_target.

(There are a number of ways that a monster can go from full health to
being dead-but-not-gibbed from a single damage infliction.  Perhaps the
most common is for the monster to take just the right amount of splash
damage from an explosive or quad explosive.)

There was a similar issue in some of the monsters' th_pain() functions,
where it was possible for the function to return without calling
monster_pain_use(), even when it should.

In theory, these issues could be fixed by rearranging the code in the
th_die() and th_pain() functions so that monster_pain_use() always gets
called when it should.

But it seems to me there's a better solution.

Instead of having the code involving pain_threshold and
monster_pain_use() in all of the different monsters' th_pain() and
th_die() functions, it would be simpler if there was just one single,
central place where pain_threshold is checked and where
monster_pain_use() is called.  The place to do this is in T_Damage().

This commit implements this solution, and removes the old code from all
of the th_pain() and th_die() functions, as it's no longer needed.

This should make it so that pain_target gets fired reliably in all
cases.