Fix the item respawn sound in deathmatch mode
From reading the progs_dump manual, it sounds like the original "item
respawn" sound effect is still supposed to be the norm in deathmatch
mode, while in single-player/coop the mapper gets to choose between two
different effects (depending on whether they set the "Respawn DM Style"
spawnflag).

However, deathmatch mode was being treated the same as
single-player/coop, meaning that items defaulted to using the "teleport"
sound and particle effects in deathmatch.

This commit makes it so that when an item respawns in deathmatch, it
always uses the original "item respawn" sound effect.