Detect whether a func_mapjamx map is being played
This commit lays the groundwork for the next bugfix I'm going to
implement.  In order to maintain backwards-compatibility with the
func_mapjamx maps, the next bugfix will need to be able to check whether
a func_mapjamx map is being played.

This commit implements a new function, DetectKnownRelease(), which is
called once, on map start.  This function checks whether a func_mapjamx
map is being played based on the name of the map (and, in the case of
the ambiguously-named "start" map, based on world.message).

DetectKnownRelease() sets a new global variable, known_release.  Other
parts of the code will be able to check this variable to determine
whether a func_mapjamx map is being played, for example:

    if (known_release == KNOWN_RELEASE_FUNC_MAPJAMX)
    {
        // Do backwards-compatibility stuff.
    }

Currently, DetectKnownRelease() only checks whether or not a map from
func_mapjamx is being played, because that's all it needs to do at
present.  However, I've deliberately written this code so that it can
easily be extended to detect maps from other releases, in case the need
arises.

The next commit will use the known_release variable to fix a bug in
a backwards-compatible way.