Fix item_health collision with BSP model entities
Normally, items can be pushed and generally moved around by entities
that have MOVETYPE_PUSH (e.g. func_door, func_plat, func_train).  For
example:

- If an item is on top of a door/plat/train, the item should move along
  with that entity.

- If an item is in front of a moving door/train, that entity should push
  the item along in front of it.

However, I noticed that item_health exhibited the following behavior
with progs_dump:

- If an item_health was on top of a door/plat/train, the item_health
  would be left floating in mid-air when that entity moved away.

- If an item_health was in front of a moving door/train, that entity
  would clip through the item_health.

The cause was a problematic hack which progs_dump inherited from the RMQ
code.  The RMQ PlaceItem() function contained a hack which was applied
to item_health, which minimized the size of the item's bounding box
before droptofloor() was called, and afterwards restored the item's
bounding box and set its movetype to MOVETYPE_NONE.  I suspect the RMQ
code did this so that item_health could be put in tight spaces without
"falling out of the level".

The problem is that this hack completely broke item_health's collision
with entities that have MOVETYPE_PUSH (doors, plats, trains, etc.).
Items usually have MOVETYPE_TOSS; changing an item_health's movetype to
MOVETYPE_NONE meant that a MOVETYPE_PUSH entity would ignore it,
resulting in the buggy behavior described above.

It seems to me that progs_dump wasn't supposed to have this change in
behavior, and only included the hack in PlaceItem() because that code
came from RMQ.  So, my first thought was to simply remove the
item_health hack.

However, it turns out that removing the hack entirely would affect
backwards-compatibility with the func_mapjamx maps.  Those maps were
created with this hack in effect, and some of them contain health boxes
which are accidentally touching walls.  These health boxes were
prevented from "falling out of the level" by this hack, so removing the
hack would cause them to "fall out of the level".

The solution that this commit implements is to leave the item_health
hack in, but to only run that code if a map from func_mapjamx is being
played.  This way, backwards-compatibility with the func_mapjamx maps is
maintained, but the buggy item_health behavior will no longer affect
other maps.

Side note: if a mapper wants to put an item_health (or indeed any item)
in a tight space without it "falling out of the level", this can already
be achieved by setting the "Suspended in Air" spawnflag.