Allow "wait" to be set for elevators
The original elevator code from Dissolution of Eternity had a hard-coded
two second wait period before an elevator could be activated again after
it started or stopped.  This meant that if the player pushed an elevator
button within two seconds of the elevator reaching a new floor, the
button wouldn't work.

This commit makes it so that the mapper can configure the wait period by
setting the elevator's "wait" field, which defaults to zero (i.e. no
wait period).