Prevent an elevator reactivating while moving
The original elevator code from Dissolution of Eternity allowed an
elevator to be activated again when it was in the process of moving
between floors (assuming the wait period was over).  This could happen
if e.g. a player on an elevator touched an elevator button at a stop
while the elevator was still moving towards that stop.

However, this reactivation sometimes resulted in seemingly buggy
behavior, e.g. the elevator moving to a different floor to what the
player would expect.

When the wait-before-reactivation was hard-coded as two seconds, this
would only have mattered in a setup where the elevator was moving for
more than two seconds between stops.  However, now that we've made the
wait period configurable via self.wait, it's easier to create a setup
where it's possible for the player to trigger this behavior.  So,
I thought something should be done about this.

It seems to me that it isn't desirable for an elevator to reactivate
when it's already in motion, so this commit simply prevents that from
happening.

In order to achieve this, this commit tightens up the handling of the
elevator's self.elevatorOnFloor and self.elevatorToFloor fields so that
they can be used to determine whether the elevator has already reached
its stop.

Note that this commit does not stop the elevator reversing direction if
it gets blocked (e.g. by the player's body); that's a different part of
the code.