Add new spawnflags to all entities (commit 4 of 4)
This commit adds the new spawnflags to the FGD files.  It also changes
the FGD files' description of spawnflag 1024 from "Not on Hard" to "Not
on Hard or Nightmare".  This is more accurate, and I thought it would
help make clear the difference between this spawnflag and the new "Not
on Hard Only" and "Not on Nightmare Only" spawnflags.