Add new spawnflags to all entities (commit 3 of 4)
This commit is the result of the big scripted edit to add a call to
SUB_Inhibit to the top of every spawn function.

I went over the results afterwards, both to check that my script hadn't
done anything it shouldn't have, and to manually tweak the indentation
of the new lines in functions that are indented differently to the
original Quake code.

The original Quake code is indented with tabs, so I defaulted to
indenting the newly-added lines with tabs, too.  However, progs_dump
contains code from a number of sources, and some functions are indented
differently.  For functions which are indented with spaces, I manually
tweaked the new lines so that they are indented with the same number of
spaces.  For functions which are indented with a mixture of tabs and
spaces, I just left the new lines indented with tabs.