Make preparations for adding the new spawnflags
To implement the new spawnflags for all entities ("Not in Coop" etc.),
a new function call will need to be added to the top of every spawn
function.  It would be very tedious to add these function calls by hand,
so I intend to add them via a scripted edit that will automate as much
of the work as possible.

The challenge with a scripted approach is making it correctly identify
which functions are spawn functions and which aren't.  The QuakeC
language itself does not have any special syntax to differentiate spawn
functions from other functions.  However, in the QuakeC source code for
the original game, spawn functions were marked by special QUAKED
comments, which contained data for the original map editor.  So, my
intention was to make my script scan the source code for these QUAKED
comments, and then have it add the new function call to the top of every
function which had a matching QUAKED comment.

This strategy might have worked as-is for the original QuakeC source
code.  However, because progs_dump contains pieces of code that
originated in various different projects, it presented some further
difficulties:

- A small number of the new QUAKED comments didn't match the format of
  the QUAKED comments from the original game.  This would have got in
  the way of automatic detection.

- A number of the new spawn functions didn't have QUAKED comments at
  all.

So, to prepare for the big scripted edit, this commit makes the
following changes:

- The QUAKED comments which were formatted inconsistently have been
  tweaked to make them consistent.

- QUAKED comments have been added for all spawn functions which lacked
  them.  I determined what the new spawn functions were by looking
  through the FGD files and the manual.  I added these new QUAKED
  comments by hand, and they are very "bare bones", containing only the
  minimum amount of information necessary for the format.  In
  particular, I didn't attempt to add descriptions of entity behavior,
  spawnflags, etc.

- This commit makes one final tweak intended to help the scripted edit
  run smoothly, which is to delete various unnecessary blank lines
  around the tops of some spawn functions (either between the function
  header and the opening brace, or between the opening brace and the
  first statement).

The actual scripted edit will be done in a later commit.