Add new spawnflags to all entities (commit 1 of 4)
This commit adds a new file, newflags.qc, which defines all of the
functions related to the new spawnflags that are going to be added to
all entities ("Not in Coop", etc.).  To implement the new spawnflags,
calls to these new functions will need to be added to various places in
the existing code.  This will be done in follow-up commits.