Fix messages not displaying with sight_trigger
Previously, when a monster with sight_trigger fired its targets on
seeing the player, if one of the targeted entities had a message, that
message would not necessarily be displayed to the player.

The reason for this was that the sight_trigger code in FoundTarget() did
not set the global "activator" variable before it called
SUB_UseTargets().  In another recent commit, I changed the call to
SUB_UseTargets() into a call to SUB_UseAndForgetTargets(), but the same
applies: "activator" needs to be set first, because this is the start of
a new target chain.

This commit fixes this by setting "activator" to self.enemy.