Fix sight_trigger firing targets multiple times
From the description in the progs_dump manual, a monster that has
sight_trigger should fire its targets when it sees the player instead of
firing its targets when it dies.  The code had two issues in relation to
this:

- A monster with sight_trigger would actually fire its targets both when
  it saw the player and when it died.  This was because SUB_UseTargets()
  would be called both in FoundTarget() and in monster_death_use().

- A monster with sight_trigger would fire its targets every time it got
  angry at something.  The mapper would have no control over how many
  times that would be.

This commit makes it so that a monster with sight_trigger will fire its
targets once and only once.