Prevent keys firing targets multiple times in Coop
Similar to weapon items, the code for the original game made key items
behave differently in Coop with regard to firing their targets.  In
Single Player, a key item would fire its targets once and only once, at
the time the player collected it.  However, in Coop, a key item would
stick around so that multiple players could each collect a copy of the
same key, and would fire its targets each time a player took a copy of
the key.

As with weapon items, this meant that a map could contain a target setup
involving a key item which would work in Single Player but which would
break in Coop.

This commit removes this inconsistency by making it so that a key item
will always fire its targets once and only once, the first time it is
collected, regardless of game mode.