Make weapons fire their targets in Coop and DM2
The QuakeC code for the original game had a quirk (arguably a bug)
whereby a weapon item would never fire its targets in Coop or
"Deathmatch 2".  This was the case because it would only fire its
targets when it disappeared, but it would never disappear in Coop or
"Deathmatch 2", because in those game modes, a weapon item sticks around
so that multiple players can collect it.

Unfortunately, this quirk meant that it was possible for a map to
include a target setup which would work in Single Player but not in
Coop.  Various maps made for vanilla Quake have this problem.  It
affects at least one of the demo maps that comes with progs_dump
(pd_ionous, which includes a weapon_lightning which fires an important
target).

This commit makes it so that in Coop or "Deathmatch 2", a weapon item
will fire its targets the first time it is touched by one of the
players.  This fixes pd_ionous in Coop, and should prevent the same
problem from affecting other maps.