Fix players sliding on a func_bob with "style 1"
Previously, if a player attempted to stand on top of a func_bob with
"style 1" which had a BSP model, before it was triggered and started
moving, the player would slide around, and would be unable to jump, as
if they were on top of a monster.

This happened because the code in func_bob_off() which set self.solid
was inconsistent with the code in func_bob_on().  func_bob_on() checked
whether self.bsporigin was true and, if so, set self.solid to SOLID_BSP.
In contrast, func_bob_off() didn't check self.bsporigin, and set
self.solid to SOLID_BBOX even when the entity had a BSP model.

This commit makes it so that func_bob_off() checks self.bsporigin and
sets self.solid to SOLID_BSP if appropriate.

In order for this to work, another change was necessary, which is that
a func_bob no longer has MOVETYPE_NONE while it's waiting to be
triggered.  This is because the combination of MOVETYPE_NONE and
SOLID_BSP would cause the engine to quit with an error.  A func_bob now
has the same movetype both before and after it starts moving: if it has
a BSP model, it has MOVETYPE_PUSH, otherwise it has MOVETYPE_FLY.
Generally speaking, this change shouldn't be noticeable because
a func_bob that's waiting to be triggered isn't moving anyway.

Finally, this commit also tweaks the code that sets self.movetype and
self.solid in func_bob_on(), both to make it more concise and to make it
mirror the new code in func_bob_off().