Fix certain func_bob spawnflag combinations
func_bob has two spawnflags which control its solidity: BOB_COLLISION
and BOB_NONSOLID.  There was a check at the top of the func_bob spawn
function which was clearly supposed to set BOB_COLLISION as a default.
However, this code had the same issue as the similar code in the
weapon_shotgun spawn function (which was fixed in commit 0eaf6eb): it
would only set BOB_COLLISION if no spawnflags were set.  However, that
didn't handle the case where BOB_COLLISION and BOB_NONSOLID weren't set
but some other spawnflag was.  In that case, the entity would end up
with neither BOB_COLLISION nor BOB_NONSOLID, and would malfunction,
because other parts of the func_bob code assume that exactly one of
these flags is set.

There was also another, related issue, which was that the code did not
prevent the mapper from setting both BOB_COLLISION and BOB_NONSOLID,
which would also cause the entity to malfunction for the same reason.

To fix both of these issues, this commit makes it so that BOB_COLLISION
will always be set, unless the mapper has set BOB_NONSOLID, in which
case BOB_COLLISION will be cleared.