Fix func_bob with "style 1" lurching when started
Previously, if a func_bob was configured to not start moving until
triggered (i.e. if it had "style 1"), then it would sometimes lurch
violently during its first movement, with the result that it would be
permanently moved away from where it was supposed to be.

This behavior appears to have been a consequence of how the entity's
think function was handled.  A func_bob's self.think is
func_bob_timer(), which controls its movement, and a func_bob would
always start thinking immediately after it was spawned.  This made sense
for a func_bob without "style 1", because it should start moving
immediately.  However, this meant that a func_bob with "style 1" would
already be thinking before it was actually able to move.  This seems to
have been the cause of the problem.

This commit makes it so that a func_bob will only start thinking when it
is allowed to move, which seems correct, and eliminates the "lurching"
problem.

This commit includes two other tweaks to support this:

- The first assignment to self.nextthink now takes into account whether
  the func_bob has a BSP model.  If it does, self.nextthink is set
  relative to self.ltime, and if it doesn't, self.nextthink is set
  relative to "time".

- Previously, if a func_bob was triggered multiple times, func_bob_on()
  would be called multiple times.  This seemed inappropriate, and could
  potentially cause problems now that self.nextthink is set in
  func_bob_on().  So, func_bob_on() now sets self.use to SUB_Null() so
  that repeated triggerings will have no effect.