Fix messages not displaying with func_counter
This commit fixes a bug which affected both func_counter and its
companion entity func_oncount.  The code for these entities was
originally from the Hipnotic mission pack (Scourge of Armagon), and the
bug was in the original code.

The problem was that these entities did not handle the global
"activator" variable correctly.  As a result, entities targeted by these
entities would not display their messages to the player.

The purpose of the "activator" variable is to keep track of the entity
that caused a target chain to activate.  Generally speaking, the first
entity in a target chain is responsible for setting "activator" to the
entity that activated it (e.g. the player) before it calls
SUB_UseTargets().  The reason this affects the display of messages is
that SUB_UseTargets() prints any messages to the "activator" entity.
Note that this applies to any calls to SUB_UseTargets() made by any of
the entities in the target chain, so if an entity ever sets "activator"
to the wrong thing, it will affect entities further down the chain.

In the usual case where a func_counter is the target of another entity,
"activator" should already have been set to the correct entity (e.g. the
player) before the func_counter's self.use function is called.
A func_counter can have different functions set as its self.use
function, but the one that's relevant to this issue is
counter_off_use().  When it's called, counter_off_use() does one of two
things:

- If a "delay" value was set for the func_counter, then
  counter_off_use() schedules counter_think() to be called as the
  entity's next think function.

- If no "delay" value was set, then counter_off_use() just calls
  counter_think() directly.

It's counter_think() that makes the call to SUB_UseTargets() which
continues the target chain.  In the original code, counter_think() set
"activator" to "other" before calling SUB_UseTargets(), but that wasn't
right, because whichever way counter_think() was called, "other" wasn't
the entity that activated the target chain.

This commit makes it so that counter_off_use() stores the value of
"activator" in the self.enemy field, and counter_think() then copies
this value back into "activator" before calling SUB_UseTargets().  This
should ensure that "activator" gets set to the right thing regardless of
whether counter_think() is called as a think function or called directly
by counter_off_use().

The above solution fixes the handling of "activator" in the common case
where a func_counter is fired by another entity.  However,
a func_counter that has the COUNTER_START_ON spawnflag isn't fired by
another entity.  In that case, what the original code did was call
counter_off_use() as the entity's first think function instead of as
a self.use function.  But, because counter_off_use() was being called as
a think function, there was no guarantee what entity "activator" would
refer to at the time the function was called.

To handle COUNTER_START_ON, this commit adds a new function,
counter_start_on_think(), which will be called as the first think
function instead of counter_off_use().  counter_start_on_think() simply
sets "activator" to "world" and then calls counter_off_use().  This
ensures that "activator" isn't some random entity.

Note that this means that entities targeted by a func_counter with
COUNTER_START_ON still won't display messages to any of the players, but
that seems correct, because none of the players caused the activation.

Finally, this commit also fixes func_oncount.  The original code had the
same problem as the func_counter code, in that it always set "activator"
to "other" before calling SUB_UseTargets().  In the case of
func_oncount, SUB_UseTargets() is called from oncount_use().
oncount_use() is always called as a self.use function, therefore
"activator" should already be correct when this function is called.  So,
this commit just removes the line that set "activator" to "other".