Fix "delay" making func_counter not work
This commit fixes another bug in the code for func_counter (from Scourge
of Armagon), which was that a func_counter with a non-zero "delay" value
would fail to activate any func_oncount entities that it targeted.

For a func_counter, "delay" is supposed to be an initial delay before
the entity starts firing its targets after it is activated.  The
original code had two issues:

- The first issue was that this use of "delay" conflicted with the
  normal use of "delay" which is implemented in the SUB_UseTargets()
  function.  When SUB_UseTargets() is called, if self.delay is non-zero,
  then a DelayedUse entity will be spawned to fire the target after the
  specified delay.  So, a "delay" for func_counter would effectively be
  applied twice: once in the func_counter code before the entity started
  firing its targets, and once in SUB_UseTargets() each time the targets
  were fired.

- The second issue follows on from the first.  When a func_oncount is
  targeted, it makes a decision based on two things: whether the entity
  that fired it has the classname "counter", and whether that entity's
  "state" value matches its own "count" value.  In the usual case where
  a func_oncount was fired by a func_counter, this worked as expected.
  However, this did not work when the func_counter had a "delay" set,
  because in that case it wasn't the func_counter that fired the
  func_oncount: it was the DelayedUse entity spawned by
  SUB_UseTargets()!  The result was that the func_oncount wouldn't do
  anything.

To fix both of these issues, this commit implements the following
solution:

- When a func_counter is spawned, its "delay" value is copied into
  a different field, "pausetime", and "delay" is then set to zero.

- The func_counter code then checks "pausetime" instead of "delay".

Because "delay" is set to zero, the delay won't be applied a second time
in SUB_UseTargets(), which eliminates the first issue, and
SUB_UseTargets() won't spawn a DelayedUse entity, which eliminates the
second issue.